openapi updates
